Refactored TestMergingDatabaseDataSetConsumer to use URLs in place of files. This gives it the capability, for example, to work also with files/folders inside a Jar.
